Refactor chat bubbles into a separate ChatWindow object
