+++ src/main/java/Mia.java
@@ -1,4 +1,3 @@
-import java.util.Arrays;
 import java.util.Scanner;
 
 public class Mia {
@@ -9,9 +8,9 @@ public class Mia {
 
     public void run() {
         System.out.println(logo);
+        final ChatWindow window = new ChatWindow(50);
 
-        final int windowWidth = 50;
-        printResponse(new Span("Hello there!"), windowWidth);
+        window.printResponse(new Span("Hello there!"));
         final Scanner sc = new Scanner(System.in);
 
         System.out.print("Enter a command: ");
@@ -20,61 +19,61 @@ public class Mia {
 
             // Replaces the entered command (previous line) with a bubble
             System.out.print("\u001B[1A\u001B[K");
-            printCommand(new Span(line));
+            window.printCommand(new Span(line));
 
             if (line.equals("bye")) {
-                printResponse(new Span("See you!"), windowWidth);
+                window.printResponse(new Span("See you!"));
                 break;
             } else if (line.equals("list")) {
-                printResponse(new Span(tasksManager.toString()), windowWidth);
+                window.printResponse(new Span(tasksManager.toString()));
                 System.out.print("Enter a command: ");
                 continue;
             } else if (line.startsWith("delete ")) {
                 final int number = Integer.parseInt(line.substring(7));
                 if (tasksManager.deleteTask(number)) {
-                    printResponse(new Span("Task has been deleted!"), windowWidth);
+                    window.printResponse(new Span("Task has been deleted!"));
                 } else {
-                    printResponse(new Span(String.format(
+                    window.printResponse(new Span(String.format(
                             "Something went wrong when deleting task %d! Likely, you specified an invalid task number.",
-                            number)), windowWidth);
+                            number)));
                 }
                 System.out.print("Enter a command: ");
                 continue;
             } else if (line.startsWith("mark ")) {
                 final int number = Integer.parseInt(line.substring(5));
                 if (tasksManager.checkTask(number)) {
-                    printResponse(new Span("Task has been marked as done!"), windowWidth);
+                    window.printResponse(new Span("Task has been marked as done!"));
                 } else {
-                    printResponse(new Span(String.format(
+                    window.printResponse(new Span(String.format(
                             "Task not modified! Either the task is already done, or you specified an invalid task number %d.",
-                            number)), windowWidth);
+                            number)));
                 }
                 System.out.print("Enter a command: ");
                 continue;
             } else if (line.startsWith("unmark ")) {
                 final int number = Integer.parseInt(line.substring(7));
                 if (tasksManager.uncheckTask(number)) {
-                    printResponse(new Span("Task has been marked as not done!"), windowWidth);
+                    window.printResponse(new Span("Task has been marked as not done!"));
                 } else {
-                    printResponse(new Span(String.format(
+                    window.printResponse(new Span(String.format(
                             "Task not modified! Either the task is still not done, or you specified an invalid task number %d.",
-                            number)), windowWidth);
+                            number)));
                 }
                 System.out.print("Enter a command: ");
                 continue;
             } else if (line.startsWith("todo ")) {
                 final Task todo = new Todo(line.substring(5));
                 tasksManager.addTask(todo);
-                printResponse(new Span(String.format("Added todo \"%s\" to tasks list!", todo.getTitle())), windowWidth);
+                window.printResponse(new Span(String.format("Added todo \"%s\" to tasks list!", todo.getTitle())));
                 continue;
             } else if (line.startsWith("deadline ")) {
                 final String[] data = line.substring(9).split("/by", 2);
                 if (data.length == 2) {
                     final Task deadline = new Deadline(data[0], data[1]);
                     tasksManager.addTask(deadline);
-                    printResponse(new Span(String.format("Added \"%s\" (task with deadline) to tasks list!", deadline.getTitle())), windowWidth);
+                    window.printResponse(new Span(String.format("Added \"%s\" (task with deadline) to tasks list!", deadline.getTitle())));
                 } else {
-                    printResponse(new Span("Incorrect format of deadline command!"), windowWidth);
+                    window.printResponse(new Span("Incorrect format of deadline command!"));
                 }
                 continue;
             } else if (line.startsWith("event ")) {
@@ -82,65 +81,16 @@ public class Mia {
                 if (data.length == 2) {
                     final Task event = new Event(data[0], data[1]);
                     tasksManager.addTask(event);
-                    printResponse(new Span(String.format("Added new event \"%s\" to tasks list!", event.getTitle())), windowWidth);
+                    window.printResponse(new Span(String.format("Added new event \"%s\" to tasks list!", event.getTitle())));
                 } else {
-                    printResponse(new Span("Incorrect format of event command!"), windowWidth);
+                    window.printResponse(new Span("Incorrect format of event command!"));
                 }
                 continue;
             }
-            printResponse(new Span("Sorry boss, I don't know what you are trying to say 😟"), windowWidth);
+            window.printResponse(new Span("Sorry boss, I don't know what you are trying to say 😟"));
             System.out.print("Enter a command: ");
         }
         sc.close();
     }
 
-    private static Text[] breakLines(Text data) {
-        // Break strings at the next word after line length hits 30 characters
-        return data.lines()
-                .flatMap((s) -> Arrays.stream(s.split("(?<=\\G\\b?.{30,}\\s)")))
-                .map(Text::strip)
-                .toArray(Text[]::new);
-    }
-
-    private static void printResponse(Text response, int windowWidth) {
-        final Text[] lines = breakLines(response);
-        int maxLength = 3; // prevents negative count
-        for (int i = 0; i < lines.length; i++) {
-            if (maxLength < lines[i].length()) {
-                maxLength = lines[i].length();
-            }
-        }
-        final int paddingLength = windowWidth - maxLength - 4;
-        StringBuilder sb = new StringBuilder(" ".repeat(paddingLength))
-                .append("╭").append("─".repeat(maxLength + 2)).append("╮\n");
-        // Pad lines right
-        final String formatString = "%-" + maxLength + "s";
-        for (int i = 0; i < lines.length; i++) {
-            sb.append(" ".repeat(paddingLength))
-                    .append("│ ").append(String.format(formatString, lines[i])).append(" │\n");
-        }
-        sb.append(" ".repeat(paddingLength))
-                .append("╰").append("─".repeat(maxLength)).append("╮┬╯\n")
-                .append(" ".repeat(windowWidth - 7)).append("MIA ╰╯ \n");
-        System.out.printf(sb.toString());
-    }
-
-    private static void printCommand(Text command) {
-        final Text[] lines = breakLines(command);
-        int maxLength = 3; // prevents negative count
-        for (int i = 0; i < lines.length; i++) {
-            if (maxLength < lines[i].length()) {
-                maxLength = lines[i].length();
-            }
-        }
-        StringBuilder sb = new StringBuilder("╭").append("─".repeat(maxLength + 2)).append("╮\n");
-        // Pad lines right
-        final String formatString = "%-" + maxLength + "s";
-        for (int i = 0; i < lines.length; i++) {
-            sb.append("│ ").append(String.format(formatString, lines[i])).append(" │\n");
-        }
-        sb.append("╰┬╭").append("─".repeat(maxLength)).append("╯\n")
-                .append(" ╰╯ You").append(" ".repeat(maxLength - 3)).append("\n");
-        System.out.printf(sb.toString());
-    }
 }
